+++ src/renderers/homeworlds.ts
@@ -1,4 +1,3 @@
-/* eslint-disable no-console */
 import { Box, Svg, Use } from "@svgdotjs/svg.js";
 import { rectOfRects } from "../grids";
 import type { IPoint } from "../grids/_base";
@@ -29,6 +28,8 @@ interface ISystem {
 export class HomeworldsRenderer extends RendererBase {
 
     public static readonly rendererName: string = "homeworlds";
+    private backColour = "#000";
+    private contrastColour = "#fff";
 
     constructor() {
         super();
@@ -53,6 +54,11 @@ export class HomeworldsRenderer extends RendererBase {
         }
         if (this.json.board.length !== this.json.pieces.length) {
             throw new Error("The `board` and `pieces` arrays must be the same length.");
+        }
+
+        if ( (this.json.options) && (this.json.options.includes("hw-light")) ) {
+            this.backColour = "#fff";
+            this.contrastColour = "#000";
         }
 
         // PIECES
@@ -207,11 +213,11 @@ export class HomeworldsRenderer extends RendererBase {
         const backHeight = maxy - miny + (vbuffer * 2);
         const backX = minx - vbuffer;
         const backY = miny - vbuffer;
-        const handler = draw.rect(backWidth, backHeight).id("_void").fill("black").opacity(0).move(backX, backY).back();
+        const handler = draw.rect(backWidth, backHeight).id("_void").fill(this.backColour).opacity(0).move(backX, backY).back();
         if (this.options.boardClick !== undefined) {
             handler.click((e : Event) => {this.options.boardClick!(-1, -1, `_void`); e.stopPropagation();});
         }
-        draw.rect(backWidth, backHeight).id("_backfill").fill("black").move(backX, backY).back();
+        draw.rect(backWidth, backHeight).id("_backfill").fill(this.backColour).move(backX, backY).back();
 
         // Place the stash
         if ( (this.json.areas === undefined) || (! Array.isArray(this.json.areas)) || (this.json.areas.length !== 1) ) {
@@ -227,15 +233,12 @@ export class HomeworldsRenderer extends RendererBase {
         if ( (! stash.hasOwnProperty("R")) || (! stash.hasOwnProperty("B")) || (! stash.hasOwnProperty("G")) || (! stash.hasOwnProperty("Y"))) {
             throw new Error("Malformed stash. The properties 'R', 'B', 'G', and 'Y' are required.");
         }
-        const stashCellSize = 50;
+        const stashCellSize = 35;
         const sgrid = rectOfRects({gridWidth: 3, gridHeight: 5, cellHeight: stashCellSize * 2, cellWidth: stashCellSize});
-        const cxBox = stashCellSize * 1.75;
-        const boxWidth = stashCellSize * 2.5;
-        const dxBox = cxBox - (boxWidth / 2);
 
         const sgroup = this.rootSvg.group().id("stash"); // .size(stashCellSize * 3, stashCellSize * 8);
-        sgroup.text("Global Stash").fill("black").center(cxBox, stashCellSize * -1);
-
+        let realcx = 0;
+        const pcgroup = sgroup.group();
         const colours = ["R", "B", "G", "Y"];
         for (let i = 0; i < colours.length; i++) {
             const colour = colours[i];
@@ -262,7 +265,7 @@ export class HomeworldsRenderer extends RendererBase {
                         throw new Error(`The glyph you requested (${ship}) does not contain the necessary information for scaling. Please use a different sheet or contact the administrator.`);
                     }
                 }
-                const use = sgroup.use(piece);
+                const use = pcgroup.use(piece);
                 if (this.options.boardClick !== undefined) {
                     use.click((e : Event) => {this.options.boardClick!(-1, -1, `${colour}${size}`); e.stopPropagation();});
                 }
@@ -275,40 +278,51 @@ export class HomeworldsRenderer extends RendererBase {
                 // Also account for the width difference to create more even column spacing
                 let evenSpacing = 0;
                 if (size === "1") {
-                    // evenSpacing = 10.9375 * (500 / 180) * factor;
+                    evenSpacing = ((stashCellSize / 2) * 2.1);
                 } else if (size === "2") {
-                    evenSpacing = 10.9375 * (500 / 180) * factor;
+                    evenSpacing = (stashCellSize / 2);
+                    realcx = point.x + evenSpacing + (stashCellSize * 0.9);
                 }
 
-                const newx = point.x - evenSpacing;
+                const newx = point.x + evenSpacing;
                 const newy = point.y - stackingOffset;
                 use.dmove(newx, newy);
                 use.scale(factor, newx, newy);
             }
+        }
+        const cxBox = realcx;
+        const boxWidth = stashCellSize * 2.5;
+        const dxBox = cxBox - (boxWidth / 2);
+
+        sgroup.text("Global Stash").fill("black").center(cxBox, stashCellSize * -1.25);
+        // Add button bar unless told not to
+        if ( (! this.json.options) || (! this.json.options.includes("hw-no-buttons")) ) {
+            const top = stashCellSize * 7.5;
+            const height = stashCellSize * 0.7;
             // Add "sacrifice" box
-            sgroup.text("Sacrifice").fill("black").center(cxBox, stashCellSize * 8.35);
-            const sacrect = sgroup.rect(boxWidth, stashCellSize * 0.7).id("_sacrificeclick").fill({opacity: 0}).stroke({color: "black", width: 1});
+            sgroup.text("Sacrifice").fill("black").center(cxBox, top + (height / 2));
+            const sacrect = sgroup.rect(boxWidth, height).id("_sacrificeclick").fill({opacity: 0}).stroke({color: "black", width: 1});
             if (this.options.boardClick !== undefined) {
                 sacrect.click(() => this.options.boardClick!(-1, -1, "_sacrifice"));
             }
-            sacrect.dmove(dxBox, stashCellSize * 8);
+            sacrect.dmove(dxBox, top);
             // Add "pass" box
-            sgroup.text("Pass").fill("black").center(cxBox, stashCellSize * 9.05);
-            const passrect = sgroup.rect(boxWidth, stashCellSize * 0.7).id("_passclick").fill({opacity: 0}).stroke({color: "black", width: 1});
+            sgroup.text("Pass").fill("black").center(cxBox, top + (height * 1.5));
+            const passrect = sgroup.rect(boxWidth, height).id("_passclick").fill({opacity: 0}).stroke({color: "black", width: 1});
             if (this.options.boardClick !== undefined) {
                 passrect.click(() => this.options.boardClick!(-1, -1, "_pass"));
             }
-            passrect.dmove(dxBox, stashCellSize * 8.7);
+            passrect.dmove(dxBox, top + height);
             // Add "catastrophe" box
-            sgroup.text("Catastrophe").fill("black").center(cxBox, stashCellSize * 9.75);
-            const catrect = sgroup.rect(boxWidth, stashCellSize * 0.7).id("_catastropheclick").fill({opacity: 0}).stroke({color: "black", width: 1});
+            sgroup.text("Catastrophe").fill("black").center(cxBox, top + (height * 2.5));
+            const catrect = sgroup.rect(boxWidth, height).id("_catastropheclick").fill({opacity: 0}).stroke({color: "black", width: 1});
             if (this.options.boardClick !== undefined) {
                 catrect.click(() => this.options.boardClick!(-1, -1, "_catastrophe"));
             }
-            catrect.dmove(dxBox, stashCellSize * 9.4);
+            catrect.dmove(dxBox, top + (height * 2));
         }
 
-        sgroup.move(minx - (sgroup.width() as number) - stashCellSize, miny - (((sgroup.height() as number) - (maxy - miny)) / 2));
+        sgroup.move(minx - (sgroup.width() as number) - (stashCellSize * 1), miny - (((sgroup.height() as number) - (maxy - miny)) / 2));
         const box = draw.bbox();
         const padding = 2;
         draw.viewbox(box.x - padding, box.y - padding, box.width + (padding * 2), box.height + (padding * 2));
@@ -417,8 +431,11 @@ export class HomeworldsRenderer extends RendererBase {
                     if (used.length > 0) {
                         const prev = used[used.length - 1];
                         const curr = use.rbox();
-                        const dist = prev.x2 - curr.x + pieceSpacing;
-                        use.dx(dist / factor);
+                        let dist = 0;
+                        if (curr.w < cellSize) {
+                            dist = prev.x2 - curr.x + pieceSpacing;
+                            use.dx(dist / factor);
+                        }
                     }
                     used.push(use.rbox());
                 }
@@ -443,8 +460,11 @@ export class HomeworldsRenderer extends RendererBase {
                         if (used.length > 0) {
                             const prev = used[used.length - 1];
                             const curr = use.rbox();
-                            const dist = prev.y2 - curr.y + pieceSpacing;
-                            use.dy(dist / factor);
+                            let dist = 0;
+                            if (curr.w < cellSize) {
+                                dist = prev.y2 - curr.y + pieceSpacing;
+                                use.dy(dist / factor);
+                            }
                         }
                         used.push(use.rbox());
                     }
@@ -467,8 +487,11 @@ export class HomeworldsRenderer extends RendererBase {
                         if (used.length > 0) {
                             const prev = used[used.length - 1];
                             const curr = use.rbox();
-                            const dist = prev.y2 - curr.y + pieceSpacing;
-                            use.dy(dist / factor);
+                            let dist = 0;
+                            if (curr.w < cellSize) {
+                                dist = prev.y2 - curr.y + pieceSpacing;
+                                use.dy(dist / factor);
+                            }
                         }
                         used.push(use.rbox());
                     }
@@ -487,15 +510,15 @@ export class HomeworldsRenderer extends RendererBase {
         nested.size(realWidth, realHeight).viewbox(realX - 2, realY - 2, realWidth + 4, realHeight + 4);
 
         // Add fill and border
-        let stroke: any = {width: 2, color: "#fff"};
+        let stroke: any = {width: 2, color: this.contrastColour};
         if (highlight !== undefined) {
             stroke = {width: 5, color: highlight, dasharray: "4"};
         }
         if (this.options.boardClick !== undefined) {
-            nested.rect(realWidth, realHeight).fill("#fff").opacity(0).dmove(realX, realY).back().click(() => this.options.boardClick!(0, 0, sys.name));
+            nested.rect(realWidth, realHeight).fill(this.contrastColour).opacity(0).dmove(realX, realY).back().click(() => this.options.boardClick!(0, 0, sys.name));
         }
         // eslint-disable-next-line @typescript-eslint/no-unsafe-argument
-        nested.rect(realWidth, realHeight).fill("black").stroke(stroke).dmove(realX, realY).back();
+        nested.rect(realWidth, realHeight).fill(this.backColour).stroke(stroke).dmove(realX, realY).back();
 
         // Add name
         // nested.text(name).move(grid[0][0].x, grid[0][0].y).fill("#fff");
@@ -507,7 +530,7 @@ export class HomeworldsRenderer extends RendererBase {
         nested.text(sysLabel)
             .font({
                 anchor: "start",
-                fill: "#fff",
+                fill: this.contrastColour,
                 size: fontsize,
             })
             .attr("alignment-baseline", "hanging")