+++ src/schemas/schema.d.ts
@@ -31,9 +31,9 @@ export interface APRenderRep {
     | "homeworlds-orig"
     | "entropy";
   /**
-   * A list of flags to pass to the renderer. `rotate-pieces` signals that the pieces must also rotate when the board rotates. It's not done by default because it's so rarely needed. The `hide-labels` option hides the external row/column labels. `no-border` hides the very outside border of the square boards.
+   * A list of flags to pass to the renderer. `rotate-pieces` signals that the pieces must also rotate when the board rotates. It's not done by default because it's so rarely needed. The `hide-labels` option hides the external row/column labels. `no-border` hides the very outside border of the square boards. The `hw-*` options are for Homeworlds. The option `clickable-edges` only applies to rect-of-hex boards and makes the individual edges clickable.
    */
-  options?: ("rotate-pieces" | "hide-labels" | "no-border" | "hw-light" | "hw-no-buttons")[];
+  options?: ("rotate-pieces" | "hide-labels" | "no-border" | "hw-light" | "hw-no-buttons" | "clickable-edges")[];
   /**
    * Map each `piece` to an actual glyph with possible options.
    */
@@ -75,7 +75,15 @@ export interface APRenderRep {
          */
         strokeOpacity?: number;
         /**
-         * Only meaningful for 'squares*' boards. Blacks out the specified cells and disables clicking. Like with `annotations`, the renderer knows nothing about a game's notation. You must provide instead the column and row numbers, which are zero-based: 0,0 is the top row, top column.
+         * Only affects rect of hex maps. Used to fill hexes with a colour.
+         */
+        hexFill?: string;
+        /**
+         * Used to add a solid block of colour behind the entire image. This should usually be left to the client, but sometimes you want the option.
+         */
+        backFill?: string;
+        /**
+         * On `squares*` boards, blacks out the specified cells and disables clicking. For hex grids, the hex simply isn't drawn. Like with `annotations`, the renderer knows nothing about a game's notation. You must provide instead the column and row numbers, which are zero-based: 0,0 is the top row, top column.
          *
          * @minItems 1
          */
@@ -206,6 +214,40 @@ export interface APRenderRep {
             }
           | {
               /**
+               * A way of drawing arbitrary lines on the board. By default, respects the stroke width, colour, and opacity of the larger board.
+               */
+              type: "line";
+              /**
+               * Expects exactly two points. Board styles where a point is the center of a space (like the `squares` board style) instead point to the top-left corner of the space. Some experimentation may be required.
+               *
+               * @minItems 2
+               * @maxItems 2
+               */
+              points: [
+                {
+                  row: number;
+                  col: number;
+                },
+                {
+                  row: number;
+                  col: number;
+                }
+              ];
+              /**
+               * The colour of the shaded area. Can be either a number (which will be interpreted as a built-in player colour) or a hexadecimal colour string.
+               */
+              colour?: PositiveInteger | Colourstrings;
+              /**
+               * 1 is fully opaque. 0 is fully transparent.
+               */
+              opacity?: number;
+              /**
+               * Stroke width of the line
+               */
+              width?: number;
+            }
+          | {
+              /**
                * This will highlight one edge of the board, indicated by compass direction. It relies on the properties`strokeWeight` and `strokeOpacity`, if given. It does not work on the `hex-odd*`, `hex-even*`, `hex-of-cir` or `hex-of-hex` boards.
                */
               type: "edge";
@@ -220,18 +262,22 @@ export interface APRenderRep {
             }
           | {
               /**
-               * Only works for the `squares*` board styles. Draws a thick line between two adjacent cells. It doesn't check adjacency, but the results will not be what you expect otherwise.
+               * Only works for the `squares*` and rect-of-hex board styles. Draws a thick line between two adjacent cells. It doesn't check adjacency, but the results will not be what you expect otherwise.
                */
               type: "fence";
               cell: {
                 row: number;
                 col: number;
               };
-              side: "N" | "E" | "S" | "W";
+              side: "N" | "NE" | "E" | "SE" | "S" | "SW" | "W" | "NW";
               /**
                * The colour of the fence. Can be either a number (which will be interpreted as a built-in player colour) or a hexadecimal colour string.
                */
               colour?: PositiveInteger | Colourstrings;
+              /**
+               * Expressed as a multiple of the base stroke width
+               */
+              width?: number;
             }
           | {
               /**
@@ -337,6 +383,14 @@ export interface APRenderRep {
          * The text that will appear at the top left of the area
          */
         label: string;
+        /**
+         * Pattern for hex colour strings
+         */
+        background?: string;
+        /**
+         * Optional. Places a coloured bar to the left of the area, used to indicate ownership.
+         */
+        ownerMark?: PositiveInteger | Colourstrings;
       }
     | {
         type: "globalStash";